+++ libs/schemas/src/commands/user.schemas.ts
@@ -10,7 +10,7 @@ export const SignIn = {
     wallet_id: z.nativeEnum(WalletId),
     session: z.string(),
     block_info: z.string().nullish(),
-    referrer_address: z.string().optional(),
+    referrer_address: z.string().nullish(),
   }),
   output: Address.extend({
     community_base: z.nativeEnum(ChainBase),
